scripts/requirements.txt
mypy==1.16.0
mypy_extensions==1.1.0
pathspec==0.12.1
python-dotenv==1.0.1
ruff==0.11.12
typing_extensions==4.13.2
